pyproject.toml
[project]
name = "piper-draw"
version = "0.1.0"
description = "Interactively draw pipe diagrams for lattice surgery."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "networkx>=3.2",
    "stim>=1.15.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
    "pytest-randomly>=4.0.1",
    "ipython>=9.12.0",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
